Global refactoring, deleted useless dependencies, now container only one
Deleted Service Locator, DryIoc container implementation, now HWDTech container is only one. Deleted accidently added in development process comments. Added new lines in files without it.
