+++ src/SpaceBattle/Base/IAngle.cs
