+++ src/SpaceBattle/Base/ICommand.cs
