+++ src/SpaceBattle/Base/IMovable.cs
