+++ src/SpaceBattle/Base/IRotatable.cs
