+++ src/SpaceBattle/Base/IStrategy.cs
