+++ src/SpaceBattle/Entities/Commands/RotateCommand.cs
@@ -1,7 +1,6 @@
 namespace SpaceBattle.Entities.Commands;
 
 using SpaceBattle.Base;
-using SpaceBattle.Base.Collections;
 using SpaceBattle.Collections;
 
 public class RotateCommand : ICommand
@@ -16,7 +15,7 @@ public class RotateCommand : ICommand
 
     public void Run()
     {
-        var cont = ServiceLocator.Locate<IContainer>("IoC");
+        var cont = new Container();
         this.Rotatable.Rotation = cont.Resolve<IList<IAngle>>("Math.IList.IAngle.Addition", this.Rotatable.Rotation, this.Rotatable.RotationSpeed);
     }
 }
