+++ src/SpaceBattle/Entities/Commands/StartMoveCommand.cs
@@ -1,6 +1,5 @@
 namespace SpaceBattle.Entities.Commands;
 using SpaceBattle.Base;
-using SpaceBattle.Base.Collections;
 using SpaceBattle.Collections;
 
 public class StartMoveCommand : ICommand
@@ -14,7 +13,7 @@ public class StartMoveCommand : ICommand
 
     public void Run()
     {
-        var container = ServiceLocator.Locate<IContainer>("IoC");
+        var container = new Container();
         container.Resolve<ICommand>("Object.SetupProperty", mcs.UObject, "velocity", mcs.Velocity).Run();
 
         var Movable = container.Resolve<IMovable>("Entities.Adapters.IMovable", mcs);
