+++ src/SpaceBattleTests/Collections/Container_tests.cs
@@ -3,11 +3,11 @@ namespace SpaceBattleTests.Collections;
 using SpaceBattle.Collections;
 using SpaceBattle.Base;
 
-public class HWDTechContainerTests {
+public class ContainerTests {
     
     [Fact]
     void ContainerResolveRegister_Succesful() {
-        var container = new HWDTechContainer();
+        var container = new Container();
         
         container.Resolve<ICommand>("Scopes.Current.Set", container.Resolve<object>("Scopes.New", container.Resolve<object>("Scopes.Root"))).Run();
         container.Resolve<ICommand>("IoC.Register", "Test.Dependency", typeof(TestInjector)).Run();
@@ -17,4 +17,12 @@ public class HWDTechContainerTests {
         Assert.IsType(typeof(TestDependency), dep);
     }
 
+}
+
+internal class TestInjector : IStrategy
+{
+    public object Run(params object[] argv)
+    {
+        return new TestDependency();
+    }
 }
