--- src/SpaceBattleTests/Misc/Strategies/InjectContainerStrategy.cs
@@ -1,12 +1,0 @@
-namespace SpaceBattleTests.Misc.Strategies;
-using SpaceBattle.Collections;
-using SpaceBattle.Base;
-
-public class InjectContainerStrategy<ContainerType> : IStrategy
-{
-    ContainerType Container = Activator.CreateInstance<ContainerType>();
-    public object Run(params object[] argv)
-    {
-        return Container!;
-    }
-}
